backend/pyproject.toml
[tool.poetry]
name = "po-pdf-splitter"
version = "0.1.0"
description = ""
authors = ["Zach Myers <zachmyers@woosterbrush.com>"]

[tool.poetry.dependencies]
python = ">3.8"
python-dotenv = "^0.15.0"
dnspython = "^2.1.0"
fastapi = "^0.63.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
motor = "^2.3.1"
websockets = "^8.1"
python-multipart = "^0.0.5"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
flake8 = "^3.8.4"
bump2version = "^1.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"